+++ unique_toolkit/tests/language_model/test_language_model_schemas.py
@@ -1,7 +1,7 @@
 import json
 
 import pytest
-from pydantic import ValidationError
+from pydantic import BaseModel, Field, ValidationError
 
 from unique_toolkit.language_model.schemas import (
     LanguageModelAssistantMessage,
@@ -76,7 +76,7 @@ class TestLanguageModelSchemas:
                     "index": 0,
                     "finishReason": "finished",
                     "message": {
-                        "role": "system",
+                        "role": "assistant",
                         "content": "content",
                         "name": "name",
                         "toolCalls": [
@@ -100,7 +100,7 @@ class TestLanguageModelSchemas:
         choice = response.choices[0]
         assert choice.index == 0
         assert choice.finish_reason == "finished"
-        assert choice.message.role == LanguageModelMessageRole.SYSTEM
+        assert choice.message.role == LanguageModelMessageRole.ASSISTANT
         assert choice.message.content == "content"
 
         # This is temporary this code would be deleted soon.
@@ -334,3 +334,21 @@ def test_language_model_tool_name_pattern_raised_validation_error():
         )
 
         assert tool.name == name
+
+
+def test_language_model_tool_parameters_dump_from_pydantic():
+    class TestParameters(BaseModel):
+        param: str = Field(description="A parameter")
+        param2: int = Field(description="A parameter")
+
+    tool = LanguageModelTool(
+        name="DocumentSummarizerV2",
+        description="Invalid tool name",
+        parameters=TestParameters.model_json_schema(),
+    )
+
+    assert tool.parameters.properties["param"].type == "string"
+    assert tool.parameters.properties["param"].description == "A parameter"
+    assert tool.parameters.properties["param2"].type == "integer"
+    assert tool.parameters.properties["param2"].description == "A parameter"
+    assert tool.parameters.required == ["param", "param2"]