+++ unique_toolkit/CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - remove the `Event` dependency from `LanguageModelService`
 - move SDK calls to `unique_toolkit.language_model.functions`
 
+## [0.5.48] - 2025-01-19
+- Added the possibility define tool parameters with a json schema (Useful when generating tool parameters from a pydantic object)
+
+## [0.5.47] - 2025-01-07
+- Added a message builder to build language model messages conveniently without importing all different messages.
+- Move tool_calls to assistant message as not needed anywhere else.
+
 ## [0.5.46] - 2025-01-07
  - Added `AZURE_GPT_35_TURBO_0125` as new model into toolkit.